+++ routes/index.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const listRouter = require("./listRouter");
+
 const cartRouter = require("./cartRouter");
 const productRouter = require("./productRouter");
 const userRouter = require("../routes/userRouter");
@@ -7,7 +7,6 @@ const categoryRouter = require("./categoryRouter");
 
 const router = express.Router();
 
-router.use("/products", listRouter.router);
 router.use("/carts", cartRouter.router);
 router.use("/users", userRouter.router);
 router.use("/products", productRouter.router);
