+++ routes/productRouter.js
@@ -1,8 +1,11 @@
 const express = require("express");
-const router = express.Router();
 
 const productController = require("../controllers/productController");
+
+const router = express.Router();
+
 router.get("/:productId", productController.getProductDetail);
+router.get("", productController.getProductList);
 
 module.exports = {
   router,
