+++ services/productService.js
@@ -5,6 +5,30 @@ const getProductDetail = async (productId) => {
   return getProductDetail;
 };
 
+const getProductList = async (
+  categoryId,
+  productId,
+  minPrice,
+  maxPrice,
+  subcategoryId,
+  orderBy,
+  limit,
+  offset
+) => {
+  const getList = await productDao.getProductList(
+    categoryId,
+    productId,
+    minPrice,
+    maxPrice,
+    subcategoryId,
+    orderBy,
+    limit,
+    offset
+  );
+  return getList;
+};
+
 module.exports = {
   getProductDetail,
+  getProductList,
 };
